llm-finetuning-system/requirements.txt
modal>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0
httpx>=0.25.0
responses>=0.23.0
freezegun>=1.2.0
factory-boy>=3.3.0
faker>=19.0.0
locust>=2.17.0
pytest-benchmark>=4.0.0
coverage>=7.3.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
slowapi>=0.1.5
email-validator>=2.0.0
python-dotenv>=1.0.0

